refac(client):put types into seperate files and rename some files that don't make sense
