+++ client/src/components/DocumentTable/DocumentTable.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import { ClientInput } from '../ClientInput/ClientInput'
-import { Header, HeaderCell } from '../Table/TableStyles'
+import { Header, HeaderCell } from '../Table/TableComponents'
 import { type FileInfo, FileRow } from './FileRow'
 import { FileSourceSelector } from '../ClientInput/FileSourceSelector'
