+++ client/src/types/ClientInfo.ts
@@ -1,0 +1,5 @@
+export interface ClientInfo {
+    firstName: string
+    lastName: string
+    jobTitle: string
+}
