+++ client/src/types/DocumentRowState.ts
